Move custom MSSQL date types to subclass
